+++ src/api/routes/rosetta/construction.ts
@@ -514,7 +514,7 @@ export function createRosettaConstructionRouter(db: PgStore, chainId: ChainID): 
       }
       try {
         const baseTx = rawTxToBaseTx(inputTx);
-        const operations = await getOperations(baseTx, db);
+        const operations = await getOperations(baseTx, db, chainId);
         const txMemo = parseTransactionMemo(baseTx);
         let response: RosettaConstructionParseResponse;
         if (signed) {
@@ -535,6 +535,7 @@ export function createRosettaConstructionRouter(db: PgStore, chainId: ChainID): 
         res.json(response);
       } catch (error) {
         console.error(error);
+        res.status(400).json(RosettaErrors[RosettaErrorsTypes.unknownError]);
       }
     })
   );